+++ manual/source/version_history.rst
@@ -42,6 +42,8 @@ Bug Fixes
 * Clarify message when loading a field map and header variables such as "nx" and "ny" were not
   specified and therefore defaulted to 0, which is invalid. Also, complain if these are purposively
   assigned to values less than 1.
+* :code:`lhcdetailed` beam pipe now **ignores** :code:`beampipeMaterial` and uses the LHC specific
+  materials as 2K.
 
 Output Changes
 --------------